disabled protoc action fixed clean action TODO rework the whole protobuf generation stuff

+++ NaoTHSoccer/Make/tools/clean_action.lua
@@ -10,17 +10,20 @@ newaction {
 	description = "Cleanup the build and dist folders",
 
 	onStart = function()
-		print("Cleanup build and dist folders")
+		print("Cleanup build/dist folders + generated protobuf messages")
 	end,
 
 	execute = function()
-		bla = os.rmdir("../build")
-		blub = os.rmdir("../dist")
+		if os.isdir("../build") then
+			bla = os.rmdir("../build")
+		end
+		if os.isdir("../dist") then
+			blub = os.rmdir("../dist")
+		end
 		
 		-- force protobuf to recompile
-		if isfile("../Messages/.Representations.proto.lastCompile~") then
-			print("found protobuf file")
-			-- asdf = os.remove("../Messages/.Representations.proto.lastCompile~")
+		if os.isfile("../Messages/.Representations.proto.lastCompile~") then
+			asdf = os.remove("../Messages/.Representations.proto.lastCompile~")
 		end
 		--[[
 		if bla == nil then
